+++ tests/annotations/test_value.py
@@ -90,6 +90,24 @@ def test_value_annotation():
         assert v.check_against(s.lower())
         mocked_check.assert_called_with([(s.lower(), 0)])
 
+    # check custom equivalence function
+    mocked_eq = mock.MagicMock()
+    v = Value(s, equivalence_fn=mocked_eq)
+    mocked_eq.return_value = False
+    assert not v.check_against("foo")
+    mocked_eq.assert_called_with(s, "foo")
+    mocked_eq.return_value = True
+    assert v.check_against("")
+    mocked_eq.assert_called_with(s, "")
+    mocked_eq.side_effect = ValueError()
+    assert not v.check_against("")
+
+    # check for invalid return type error
+    mocked_eq.return_value = 1
+    mocked_eq.side_effect = None
+    with pytest.raises(TypeError, match=f"Custom equivalence function returned value of invalid type: {type(1)}"):
+        v.check_against(1)
+
 
 def test_attribute_annotation():
     """