+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file, and this pr
 * Fixed bug in `Value`s tracking `set`s of numbers
 * Fixed bug in timestamp comparison in `BeforeAnnotation`
 * Added `Value.check_against` and `Attribute.check_agains` per [#111](https://github.com/microsoft/pybryt/issues/111)
+* Added custom equivalence functions for value annotations per [#113](https://github.com/microsoft/pybryt/issues/113)
 
 ## 0.2.0 2021-09-01